chore: Repository의 ID 타입을 Integer에서 Long으로 변경

+++ back/src/main/java/com/back/domain/coffee/coffee/repository/CoffeeRepository.java
@@ -3,5 +3,5 @@ package com.back.domain.coffee.coffee.repository;
 import com.back.domain.delivery.delivery.entity.Delivery;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CoffeeRepository extends JpaRepository<Delivery, Integer> {
+public interface CoffeeRepository extends JpaRepository<Delivery, Long> {
 }
