+++ back/src/main/java/com/back/domain/order/order/repository/OrderRepository.java
@@ -3,5 +3,5 @@ package com.back.domain.order.order.repository;
 import com.back.domain.delivery.delivery.entity.Delivery;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface OrderRepository extends JpaRepository<Delivery, Integer> {
+public interface OrderRepository extends JpaRepository<Delivery, Long> {
 }
